ref: extract cursor handling from UserDetails to UserHsCursor
